feat: change location for styles, update index.tsx
